ccodegen: function made ready for lamdas

+++ src/semantics/ScopeGenerator.cpp
@@ -164,7 +164,7 @@ namespace semantics {
 
     void ScopeGenerator::visit(ListPattern &node) {
         if (node.type.id == TypeId::LIST) {
-            for (int i = 0; i < node.patterns.size(); ++i) {
+            for (size_t i = 0; i < node.patterns.size(); ++i) {
                 node.patterns[i]->type = node.type.subtypes[i];
                 node.patterns[i]->accept(*this);
             }
